requirements.txt
fastapi
uvicorn
flatlib
swisseph
openai
python-telegram-bot

